Refactoring login form page
