--- src/services/apiUsers.ts
@@ -1,7 +1,0 @@
-import axios from "axios";
-
-const apiUsers = axios.create({
-  baseURL: "",
-});
-
-export default apiUsers;
